Add files via upload
This repository is a valuable resource for anyone interested in learning about machine learning or predicting wine quality.

Winemakers can use the machine learning models to predict the quality of their wine before it is bottled. This can help them to identify and improve problem batches.
Wine retailers can use the machine learning models to recommend wines to their customers based on their preferences.
Wine consumers can use the machine learning models to learn more about the wines they are buying and to choose wines that are likely to be of high quality.